Database

Documentation

dbConnect :: IO Connection

The below functions are used for establishing a connection to the database

r :: (Connection -> IO a) -> IO a

setupDatabase :: IO Connection

setupDatabase initialises the database creating tables

populateMovies :: IO ()

populateMovies populates movie data into a table of its own. | Begins by preparing Connections | Then inserts and Commit