DatabaseQueries

Documentation

starringByMovieName :: String -> IO ()

Use of JOIN | Show all starring actors for a particular movie.

delByMovieName :: String -> IO ()

Use of DELETE. | Delete a record based on a given movie name

updByMovieName :: p -> String -> IO ()

Use of UPDATE. | Update a record based on a given movie name

movieByActorName :: [Char] -> IO ()

The following sql query returns the movie name based on the starring actor.

actorFrequency :: Integer -> IO ()

The following sql query returns the most frequent starring actors from the database.

dirByMovieName :: String -> IO ()

The following sql query returns the name of the director based on the movie name

ratingByMovieName :: String -> IO ()

The following sql query shows the rating based on the movie name.

rankByMovieName :: String -> IO ()

Show rank based on a movie name