SqliteToJson

Documentation

jsonFile :: FilePath

Below functions are used to read the json file

getJSON :: IO ByteString

data Movies

Constructors

Movies

Fields

rank :: Int

name :: String

director :: String

year :: Int

rating :: Double

dirFName :: String

dirLName :: String

movTableRows :: [SqlValue] -> Movies

Reads each row from the movie table

starTableRows :: [SqlValue] -> Starring

Reads each row from the starring table

convMovieToJson :: IConnection conn => conn -> IO [Movies]

Fetches all rows into Strings

convStarringToJson :: IConnection conn => conn -> IO [Starring]

generateJson :: IO ()

A do block to encode and dump the database to JSON

Orphan instances

ToJSON Starring

Instance details

Methods

toJSON :: Starring -> Value

toEncoding :: Starring -> Encoding

toJSONList :: [Starring] -> Value

toEncodingList :: [Starring] -> Encoding