Database

Documentation

dbConnect :: IO Connection

The below functions are used for establishing a connection to the database

r :: (Connection -> IO a) -> IO a

dropDatabase :: IO Connection

drops all tables in the database

movieTable :: String

Functions to specify the SQL tables

starringTable :: String

movTable_starTable :: String

setupDatabase :: IO Connection

setupDatabase initialises the database creating tables

insertOne :: String

insertTwo :: String

insertThree :: String

populateMovies :: IO ()

populateMovies populates movie data into a table of its own. | Begins by preparing Connections | Then inserts and Commit